+++ Examples/Resources/xVMNonLegacyNetworkAdapter/1-Sample_xVMNonLegacyNetworkAdapter_MultipleVM_Config.ps1
@@ -1,5 +1,5 @@
 <#PSScriptInfo
-.VERSION 1.0.2
+.VERSION 1.2.0
 .GUID 0d00c0be-77d1-4170-ae48-7ba5fdd807e7
 .AUTHOR Sergei S. Betke
 .COMPANYNAME Test-St-Petersburg
@@ -20,21 +20,21 @@
 
 <#
     .SYNOPSIS
-        Configuration that will create a legacy NICs in VM 'MyVM01'.
+        Configuration that will create a non legacy NICs in VM 'MyVM01'.
 
     .DESCRIPTION
-        Configuration that will create a legacy NICs in VM 'MyVM01'.
+        Configuration that will create a non legacy NICs in VM 'MyVM01'.
 
     .PARAMETER NodeName
         The names of one or more Hyper-V nodes to compile a configuration for.
         Defaults to 'localhost'.
 
     .EXAMPLE
-        Sample_xVMLegacyNetworkAdapter_MultipleVM
+        Sample_xVMNonLegacyNetworkAdapter_MultipleVM
 
-        Compiles a configuration that creates legacy NICs on localhost Hyper-V VM.
+        Compiles a configuration that creates non legacy NICs on localhost Hyper-V VM.
 #>
-Configuration Sample_xVMLegacyNetworkAdapter_MultipleVM_Config
+Configuration Sample_xVMNonLegacyNetworkAdapter_MultipleVM_Config
 {
     param
     (
@@ -44,11 +44,11 @@ Configuration Sample_xVMLegacyNetworkAdapter_MultipleVM_Config
     )
 
     Import-DscResource -ModuleName PSDesiredStateConfiguration
-    Import-DscResource -ModuleName xITGHyperV -Name xVMLegacyNetworkAdapter
+    Import-DscResource -ModuleName xITGHyperV -Name xVMNonLegacyNetworkAdapter
 
     node $NodeName
     {
-        xVMLegacyNetworkAdapter MyVM01NIC {
+        xVMNonLegacyNetworkAdapter MyVM01NIC {
             Id = 'MyVM01-NIC'
             Name = 'MyVM01-NIC'
             SwitchName = 'SETSwitch'
@@ -56,7 +56,7 @@ Configuration Sample_xVMLegacyNetworkAdapter_MultipleVM_Config
             Ensure = 'Present'
         }
 
-        xVMLegacyNetworkAdapter MyVM02NIC {
+        xVMNonLegacyNetworkAdapter MyVM02NIC {
             Id = 'MyVM02-NIC'
             Name = 'NetAdapter'
             SwitchName = 'SETSwitch'
@@ -64,11 +64,12 @@ Configuration Sample_xVMLegacyNetworkAdapter_MultipleVM_Config
             Ensure = 'Present'
         }
 
-        xVMLegacyNetworkAdapter MyVM03NIC {
+        xVMNonLegacyNetworkAdapter MyVM03NIC {
             Id = 'MyVM03-NIC'
             Name = 'NetAdapter'
             SwitchName = 'SETSwitch'
             VMName = 'MyVM03'
+            MacAddress = '99999999'
             Ensure = 'Present'
         }
     }