+++ Tests/Unit/ITG_xVMNonLegacyNetworkAdapter.Tests.ps1
@@ -1,6 +1,6 @@
 #region HEADER
 $script:dscModuleName = 'xITGHyperV'
-$script:dscResourceName = 'ITG_xVMLegacyNetworkAdapter'
+$script:dscResourceName = 'ITG_xVMNonLegacyNetworkAdapter'
 
 # Unit Test Template Version: 1.2.4
 $script:moduleRoot = Split-Path -Parent (Split-Path -Parent $PSScriptRoot)
@@ -66,7 +66,7 @@ try
                 }
                 It 'should call the expected mocks' {
                     Assert-MockCalled Get-VMNetworkAdapter -Exactly 1
-                    # Assert-MockCalled Get-VMNetworkAdapter -Exactly 1 -parameterFilter { $IsLegacy }
+                    # Assert-MockCalled Get-VMNetworkAdapter -Exactly 1 -parameterFilter { -not $IsLegacy }
                 }
             }
             Context 'Legacy Network Adapter exists' {
@@ -77,7 +77,7 @@ try
                         VMName                   = 'ManagementOS'
                         MacAddress               = '14FEB5C6CE98'
                         DynamicMacAddressEnabled = $False
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -95,7 +95,7 @@ try
                 }
                 It 'should call the expected mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    # Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    # Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'Legacy VM Network Adapter does not exist' {
@@ -111,7 +111,7 @@ try
                 }
                 It 'should call the expected mocks' {
                     Assert-MockCalled Get-VMNetworkAdapter -Exactly 1
-                    # Assert-MockCalled Get-VMNetworkAdapter -Exactly 1 -parameterFilter { $IsLegacy }
+                    # Assert-MockCalled Get-VMNetworkAdapter -Exactly 1 -parameterFilter { -not $IsLegacy }
                 }
             }
             Context 'Legacy VM Network Adapter exists' {
@@ -121,7 +121,7 @@ try
                         SwitchName               = 'HostSwitch'
                         VMName                   = 'VM02'
                         DynamicMacAddressEnabled = $True
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -139,10 +139,10 @@ try
                 }
                 It 'should call the expected mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    # Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    # Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
-            Context 'Get-VMNetworkAdapter return some legacy NIC' {
+            Context 'Get-VMNetworkAdapter return some non legacy NIC' {
                 Mock -CommandName Get-VMNetworkAdapter -MockWith {
                     return @(
                         [PSObject] @{
@@ -150,7 +150,7 @@ try
                             SwitchName               = 'Switch'
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $True
-                            IsLegacy                 = $True
+                            IsLegacy                 = $False
                         },
                         [PSObject] @{
                             Name                     = 'NIC3'
@@ -158,7 +158,7 @@ try
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $False
                             MacAddress               = '99999999'
-                            IsLegacy                 = $True
+                            IsLegacy                 = $False
                         }
                     )
                 }
@@ -180,7 +180,7 @@ try
                     # Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
                 }
             }
-            Context 'Get-VMNetworkAdapter return just non legacy NIC' {
+            Context 'Get-VMNetworkAdapter return just legacy NIC' {
                 Mock -CommandName Get-VMNetworkAdapter -MockWith {
                     return @(
                         [PSObject] @{
@@ -188,7 +188,7 @@ try
                             SwitchName               = 'Switch'
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $True
-                            IsLegacy                 = $False
+                            IsLegacy                 = $True
                         },
                         [PSObject] @{
                             Name                     = 'NIC3'
@@ -196,7 +196,7 @@ try
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $False
                             MacAddress               = '99999999'
-                            IsLegacy                 = $False
+                            IsLegacy                 = $True
                         }
                     )
                 }
@@ -254,8 +254,8 @@ try
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
-                    #Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
+                    #Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                 }
             }
@@ -277,8 +277,8 @@ try
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
-                    #Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
+                    #Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                 }
             }
@@ -288,7 +288,7 @@ try
                         Name       = 'NIC04'
                         SwitchName = 'Switch04'
                         VMName     = 'ManagementOS'
-                        IsLegacy   = $True
+                        IsLegacy   = $False
                     }
                 }
                 Mock Add-VMNetworkAdapter
@@ -306,7 +306,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 1
                 }
@@ -318,7 +318,7 @@ try
                         SwitchName               = 'Switch04'
                         VMName                   = 'VM04'
                         DynamicMacAddressEnabled = $True
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
                 Mock Add-VMNetworkAdapter
@@ -336,7 +336,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 1
                 }
@@ -348,7 +348,7 @@ try
                         SwitchName               = 'SwitchDifferent'
                         VMName                   = 'VM04'
                         DynamicMacAddressEnabled = $True
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
                 Mock Add-VMNetworkAdapter
@@ -367,7 +367,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Connect-VMNetworkAdapter -Exactly 1
@@ -381,7 +381,7 @@ try
                         VMName                   = 'VM04'
                         DynamicMacAddressEnabled = $False
                         MacAddress               = '99999998'
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
                 Mock Add-VMNetworkAdapter
@@ -402,7 +402,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Set-VMNetworkAdapter -Exactly 1
@@ -416,7 +416,7 @@ try
                         SwitchName               = 'SwitchDifferent'
                         VMName                   = 'VM04'
                         DynamicMacAddressEnabled = $True
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
                 Mock Add-VMNetworkAdapter
@@ -437,7 +437,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Set-VMNetworkAdapter -Exactly 1
@@ -452,7 +452,7 @@ try
                         VMName                   = 'VM04'
                         DynamicMacAddressEnabled = $False
                         MacAddress               = '99999999'
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
                 Mock Add-VMNetworkAdapter
@@ -472,7 +472,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Set-VMNetworkAdapter -Exactly 1
@@ -499,14 +499,14 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                     Assert-MockCalled -commandName Add-VMNetworkAdapter -Exactly 1
                     Assert-MockCalled -commandName Remove-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Set-VMNetworkAdapter -Exactly 0
                     Assert-MockCalled -commandName Connect-VMNetworkAdapter -Exactly 0
                 }
             }
-            Context 'Get-VMNetworkAdapter return some legacy NIC but should not exist' {
+            Context 'Get-VMNetworkAdapter return some non legacy NIC but should not exist' {
                 Mock -CommandName Get-VMNetworkAdapter -MockWith {
                     return @(
                         [PSObject] @{
@@ -514,7 +514,7 @@ try
                             SwitchName               = 'Switch'
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $True
-                            IsLegacy                 = $True
+                            IsLegacy                 = $False
                         },
                         [PSObject] @{
                             Name                     = 'NIC3'
@@ -522,7 +522,7 @@ try
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $False
                             MacAddress               = '99999999'
-                            IsLegacy                 = $True
+                            IsLegacy                 = $False
                         }
                     )
                 }
@@ -550,7 +550,7 @@ try
                     Assert-MockCalled -commandName Connect-VMNetworkAdapter -Exactly 0
                 }
             }
-            Context 'Get-VMNetworkAdapter return just non legacy NICs but should exist legacy NICs' {
+            Context 'Get-VMNetworkAdapter return just legacy NICs but should exist legacy NICs' {
                 Mock -CommandName Get-VMNetworkAdapter -MockWith {
                     return @(
                         [PSObject] @{
@@ -558,7 +558,7 @@ try
                             SwitchName               = 'Switch'
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $True
-                            IsLegacy                 = $False
+                            IsLegacy                 = $True
                         },
                         [PSObject] @{
                             Name                     = 'NIC3'
@@ -566,7 +566,7 @@ try
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $False
                             MacAddress               = '99999999'
-                            IsLegacy                 = $False
+                            IsLegacy                 = $True
                         }
                     )
                 }
@@ -626,7 +626,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter does not exist but should' {
@@ -643,7 +643,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'Adapter exists but should not exist' {
@@ -652,7 +652,7 @@ try
                         Name       = 'NIC10'
                         SwitchName = 'Switch10'
                         VMName     = 'ManagementOS'
-                        IsLegacy   = $True
+                        IsLegacy   = $False
                     }
                 }
 
@@ -667,7 +667,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter exists but should not exist' {
@@ -676,7 +676,7 @@ try
                         Name       = 'NIC10'
                         SwitchName = 'Switch10'
                         VMName     = 'VM10'
-                        IsLegacy   = $True
+                        IsLegacy   = $False
                     }
                 }
 
@@ -691,7 +691,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'Adapter exists and no action needed' {
@@ -700,7 +700,7 @@ try
                         Name       = 'NIC10'
                         SwitchName = 'Switch10'
                         VMName     = 'ManagementOS'
-                        IsLegacy   = $True
+                        IsLegacy   = $False
                     }
                 }
 
@@ -715,7 +715,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter exists and no action needed' {
@@ -725,7 +725,7 @@ try
                         SwitchName               = 'Switch10'
                         VMName                   = 'VM10'
                         DynamicMacAddressEnabled = $True
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -740,7 +740,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter exists but with dynamic MAC address' {
@@ -750,7 +750,7 @@ try
                         SwitchName               = 'Switch10'
                         VMName                   = 'VM10'
                         DynamicMacAddressEnabled = $True
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -766,7 +766,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter exists but with static MAC address' {
@@ -777,7 +777,7 @@ try
                         VMName                   = 'VM10'
                         DynamicMacAddressEnabled = $False
                         MacAddress               = '14FEB5C6CE98'
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -792,7 +792,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter exists but with different MAC address' {
@@ -803,7 +803,7 @@ try
                         VMName                   = 'VM10'
                         DynamicMacAddressEnabled = $false
                         MacAddress               = '14FEB5C6CE99'
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -819,7 +819,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter exists but with different switch' {
@@ -830,7 +830,7 @@ try
                         VMName                   = 'VM10'
                         DynamicMacAddressEnabled = $False
                         MacAddress               = '14FEB5C6CE98'
-                        IsLegacy                 = $True
+                        IsLegacy                 = $False
                     }
                 }
 
@@ -846,7 +846,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'Adapter does not exist and no action needed' {
@@ -863,7 +863,7 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
             Context 'VM Adapter does not exist and no action needed' {
@@ -880,10 +880,10 @@ try
                 }
                 It 'should call expected Mocks' {
                     Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1
-                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
+                    #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { -not $IsLegacy }
                 }
             }
-            Context 'Get-VMNetworkAdapter return some legacy NIC' {
+            Context 'Get-VMNetworkAdapter return some non legacy NIC' {
                 Mock -CommandName Get-VMNetworkAdapter -MockWith {
                     return @(
                         [PSObject] @{
@@ -891,7 +891,7 @@ try
                             SwitchName               = 'Switch'
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $True
-                            IsLegacy                 = $True
+                            IsLegacy                 = $False
                         },
                         [PSObject] @{
                             Name                     = 'NIC3'
@@ -899,7 +899,7 @@ try
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $False
                             MacAddress               = '99999999'
-                            IsLegacy                 = $True
+                            IsLegacy                 = $False
                         }
                     )
                 }
@@ -918,7 +918,7 @@ try
                     #Assert-MockCalled -commandName Get-VMNetworkAdapter -Exactly 1 -ParameterFilter { $IsLegacy }
                 }
             }
-            Context 'Get-VMNetworkAdapter return just non legacy NIC' {
+            Context 'Get-VMNetworkAdapter return just legacy NIC' {
                 Mock -CommandName Get-VMNetworkAdapter -MockWith {
                     return @(
                         [PSObject] @{
@@ -926,7 +926,7 @@ try
                             SwitchName               = 'Switch'
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $True
-                            IsLegacy                 = $False
+                            IsLegacy                 = $True
                         },
                         [PSObject] @{
                             Name                     = 'NIC3'
@@ -934,7 +934,7 @@ try
                             VMName                   = 'VM02'
                             DynamicMacAddressEnabled = $False
                             MacAddress               = '99999999'
-                            IsLegacy                 = $False
+                            IsLegacy                 = $True
                         }
                     )
                 }