+++ CHANGELOG.md
@@ -1,3 +1,44 @@
 # Change log for xITGHyperV
 
-## Unreleased
+## 1.2.4
+
+Fix bugs
+
+- Fix ExternalModuleDependencies warning when module installed or updated
+
+## 1.2.1
+
+Fix bugs
+
+- Fix xVMLegacyNetworkAdapter and xVMNonLegacyNetworkAdapter work with legacy
+ and non legacy NIC
+
+## 1.2
+
+New features
+
+- xVMNonLegacyNetworkAdapter DSC Resource
+
+## 1.1.3
+
+- Fix publishing release notes to github.com on AppVeyor master build
+
+## 1.1.2
+
+- Publish release notes to github.com on AppVeyor master build
+
+## 1.1.1
+
+- Fix readme.md (xVMComPort Number property type)
+
+## 1.1
+
+New features
+
+- xVMComPort DSC Resource
+
+## 1.0
+
+New features
+
+- xVMLegacyNetworkAdapter DSC Resource